requirements.txt
snowflake-cli
crewai
crewai-tools
snowflake-connector-python
python-dotenv
